feat: AI chef estimated prices, budget recalc, cancel order, community & chat features
- AI Chef now shows estimated prices (VNĐ) per ingredient instead of product availability
- Auto-fill shopping request with ingredient names, quantities, and estimated prices
- Add recalculate budget button (sum of item prices) in create shopping request
- Add cancel order feature for OPEN status orders with wallet refund
- Add community feed, posts, comments features
- Add order chat and conversation list
- Add shopper dashboard, shopping activity, and checklist
- Remove old product-based activities (cart, checkout, product detail, etc.)

+++ app/src/main/java/com/example/gomarket/adapter/ShopperAdapter.java
@@ -44,12 +44,12 @@ public class ShopperAdapter extends RecyclerView.Adapter<ShopperAdapter.ShopperV
 
     private Context context;
     private List<ShopperModel> shopperList;
-    private int orderId;
+    private long orderId;
     private String recipeId;
     private List<String> orderItems;
 
     public ShopperAdapter(Context context, List<ShopperModel> shopperList,
-                          int orderId, String recipeId, List<String> orderItems) {
+                          long orderId, String recipeId, List<String> orderItems) {
         this.context = context;
         this.shopperList = shopperList;
         this.orderId = orderId;
